Add new sorting option (#442)
New sorting option added to change which item is prioritized to be
repaired. This can be changed in the newly added configuration (start
the game once with the mod to generate the default one).

The configuration key is `selectionMode` and can take 2 values :

* `ABSOLUTE`: Repairs the item with the most damage first
* `PERCENTAGE` : Repairs the item with the most `damage/max damage`
ratio first (basically repairs the item with its durability bar the
lowest first)
  * ⚠️ This is the new default value